BIDDER-3684 - Add method to disable WAL file
To do this I've create a new option wal.disable defaulted to false. I
then created a blank/dummy/noop implementation of WAL which is used if
this is set to true.

This means whenever trying to perform any operation with a WAL file
nothing actually happens and when it comes to retrieving the latest
offset we will fall through to using the next best method.